backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.169.0
python-multipart==0.0.20
pydantic==2.11.4
httpx==0.28.1
